+++ apps/client/src/components/hub/chatForm/ChatFormInput.js
@@ -8,6 +8,7 @@ import DownloadManager from "../../../lib/downloadManager.js";
 import { useConfirmWindow } from "@hooks/useConfirmWindow.js";
 
 import MessageInput from "@components/hub/elements/MessageInput";
+import MagicButton from "@components/hub/elements/MagicButton.js";
 
 import {
   addMessage,
@@ -257,15 +258,16 @@ export default function ChatFormInput({ chatMessagesBlockRef, editedMessage }) {
       });
       isConfirm &&
         messagesService.sendMessageDelete(selectedCID, [eMid], "all");
-      if (editedMessage.body !== inputRef.current.value) {
-        await messagesService.sendEditMessage(editedMessage._id, {
-          body: inputRef.current.value,
-        });
-      }
-      dispatch(addExternalProps({ [selectedCID]: {} }));
-      draftService.removeDraftWithOptions(selectedCID, "edited_mid");
-      inputRef.current.value = draftService.getLastInputText(selectedCID);
+      return;
     }
+    if (editedMessage.body !== inputRef.current.value) {
+      await messagesService.sendMessageEdit(editedMessage._id, {
+        body: inputRef.current.value,
+      });
+    }
+    dispatch(addExternalProps({ [selectedCID]: {} }));
+    draftService.removeDraftWithOptions(selectedCID, "edited_mid");
+    inputRef.current.value = draftService.getLastInputText(selectedCID);
   };
 
   useEffect(() => {
@@ -304,13 +306,16 @@ export default function ChatFormInput({ chatMessagesBlockRef, editedMessage }) {
   }, [selectedConversation, participants]);
 
   return (
-    <MessageInput
-      inputTextRef={inputRef}
-      isBlockedConv={isBlockedConv}
-      isEditAction={!!editedMessage}
-      isSending={isSendMessageDisable}
-      onSubmitFunc={editedMessage ? editMessageFunc : createAndSendMessage}
-      chatMessagesBlockRef={chatMessagesBlockRef}
-    />
+    <div className="w-full flex items-end gap-[8px]">
+      <MessageInput
+        inputTextRef={inputRef}
+        isBlockedConv={isBlockedConv}
+        isEditAction={!!editedMessage}
+        isSending={isSendMessageDisable}
+        onSubmitFunc={editedMessage ? editMessageFunc : createAndSendMessage}
+        chatMessagesBlockRef={chatMessagesBlockRef}
+      />
+      <MagicButton inputTextRef={inputRef} />
+    </div>
   );
 }