+++ apps/client/src/components/modals/ResetPasswordModal.js
@@ -7,6 +7,8 @@ import autoLoginService from "@services/autoLoginService.js";
 import EmailInput from "@components/auth/elements/EmailInput.js";
 import CustomInput from "@components/auth/elements/CustomInput.js";
 
+import autoLoginService from "@services/autoLoginService.js";
+
 import { showCustomAlert } from "@utils/GeneralUtils.js";
 
 export default function ResetPasswordModal({ isOpen, onClose }) {
@@ -38,6 +40,10 @@ export default function ResetPasswordModal({ isOpen, onClose }) {
 
   const handleSendOTP = async () => {
     setLoading(true);
+    if (!data.email?.length) {
+      showCustomAlert("No email to resend token.", "warning");
+      return;
+    }
     const isSuccess = await autoLoginService.sendOtpToken(data.email?.trim());
     isSuccess && setStep(2);
     setLoading(false);
@@ -55,7 +61,7 @@ export default function ResetPasswordModal({ isOpen, onClose }) {
   };
 
   const handleResendOTP = async () => {
-    if (!data.email) {
+    if (!data.email?.length) {
       showCustomAlert("No email to resend token.", "warning");
       return;
     }