+++ apps/client/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## 1.11.0
+
+### Features
+
+- Implemented the option to edit messages
+- Implemented the option to delete messages
+- Implemented reset password feature
+- Integrated AI agent for changing tone of messages and summarizing them
+
+### Bug Fixes:
+
+- Fixed a missing `}` error
+- Fixed hidden input container
+- Blocked drag message option for web
+- Fixed focus on login form when modal reset window is open
+
 ## 1.10.0
 
 ### Features